refactor(event-runtime): split CLI commands (WM-494)
