+++ event-runtime/cli/serve.test.mjs
@@ -1,0 +1,169 @@
+import { describe, expect, test } from "bun:test";
+import { spawn, spawnSync } from "node:child_process";
+import {
+  existsSync,
+  mkdirSync,
+  mkdtempSync,
+  readFileSync,
+  rmSync,
+  writeFileSync,
+} from "node:fs";
+import os from "node:os";
+import path from "node:path";
+import { openDb } from "../lib/db.mjs";
+import {
+  CLI,
+  DEAD_PORT,
+  assertHealthyLiveServe,
+  editStampRoot,
+  exitOf,
+  killPool,
+  makeStampRoot,
+  poolSize,
+  runCli,
+  runNotifierDeliveryCase,
+  seedRun,
+  spawnSupervisor,
+  spawnWorker,
+  waitFor,
+} from "./test-helpers.mjs";
+
+describe("serve command", () => {
+  test("serve --watch re-execs under bun --watch and binds", async () => {
+    const home = mkdtempSync(path.join(os.tmpdir(), "evrt-watch-"));
+    const port = String(59000 + (process.pid % 800));
+    const child = spawn("bun", [CLI, "serve", "--watch", "--port", port], {
+      env: { ...process.env, FACTORY_EVENT_HOME: home },
+      stdio: ["ignore", "pipe", "pipe"],
+    });
+    let out = "";
+    child.stdout.on("data", (b) => {
+      out += b;
+    });
+    child.stderr.on("data", (b) => {
+      out += b;
+    });
+    const deadline = Date.now() + 8000;
+    while (Date.now() < deadline && !out.includes("control API on")) {
+      await Bun.sleep(100);
+    }
+    child.kill("SIGTERM");
+    await new Promise((resolve) => child.once("exit", resolve));
+    expect(out).toContain(
+      "serve --watch: restarting on event-runtime/ changes",
+    );
+    expect(out).toContain("control API on");
+  });
+
+  test("serve binds the control API, starts the loop, and answers /health", async () => {
+    const home = mkdtempSync(path.join(os.tmpdir(), "evrt-serve-"));
+    const port = String(59800 + (process.pid % 100));
+    const child = spawn("bun", [CLI, "serve", "--port", port], {
+      env: { ...process.env, FACTORY_EVENT_HOME: home },
+      stdio: ["ignore", "pipe", "pipe"],
+    });
+    let out = "";
+    child.stdout.on("data", (b) => {
+      out += b;
+    });
+    child.stderr.on("data", (b) => {
+      out += b;
+    });
+    const deadline = Date.now() + 8000;
+    while (Date.now() < deadline && !out.includes("control API on")) {
+      await Bun.sleep(100);
+    }
+    let health;
+    try {
+      expect(out).toContain("control API on");
+      const res = await fetch(`http://127.0.0.1:${port}/health`);
+      expect(res.ok).toBe(true);
+      health = await res.json();
+    } finally {
+      child.kill("SIGTERM");
+      await new Promise((resolve) => child.once("exit", resolve));
+    }
+    expect(health.ok).toBe(true);
+  });
+
+  test("serve --adapter-override pi is accepted at the serve call site (OPS-517)", async () => {
+    const home = mkdtempSync(path.join(os.tmpdir(), "evrt-serve-pi-"));
+    const port = String(59800 + (process.pid % 150));
+    const child = spawn(
+      "bun",
+      [CLI, "serve", "--adapter-override", "pi", "--port", port],
+      {
+        env: { ...process.env, FACTORY_EVENT_HOME: home },
+        stdio: ["ignore", "pipe", "pipe"],
+      },
+    );
+    let out = "";
+    child.stdout.on("data", (b) => {
+      out += b;
+    });
+    child.stderr.on("data", (b) => {
+      out += b;
+    });
+    const deadline = Date.now() + 8000;
+    while (Date.now() < deadline && !out.includes("control API on")) {
+      await Bun.sleep(100);
+    }
+    child.kill("SIGTERM");
+    await new Promise((resolve) => child.once("exit", resolve));
+
+    expect(out).not.toContain('unknown --adapter-override "pi"');
+    expect(out).toContain('adapter override: all new run specs use "pi"');
+    expect(out).toContain("control API on");
+  });
+
+  test("tick runs notify as an isolated subsystem (WM-65): a throwing notifier step cannot break the tick", async () => {
+    const { tick, TICK_SUBSYSTEMS } = await import("../cli.mjs");
+    const { loadRegistry } = await import("../lib/registry.mjs");
+    expect(TICK_SUBSYSTEMS).toContain("notify");
+
+    const db = openDb(":memory:");
+    const logs = [];
+    let chainsRan = false;
+    await tick({
+      db,
+      registry: loadRegistry(),
+      policyVersion: "git:test",
+      log: (l) => logs.push(l),
+      subsystems: {
+        notify: () => {
+          throw new Error("notifier exploded");
+        },
+        chains: () => {
+          chainsRan = true;
+        },
+      },
+    });
+    expect(logs.some((l) => l.includes("tick notify: notifier exploded"))).toBe(
+      true,
+    );
+    expect(chainsRan).toBe(true);
+  });
+
+  test("tick with FACTORY_EVENT_NOTIFY=1 pushes a human_needed park through the stub notifier exactly once", async () => {
+    const { delivery } = await runNotifierDeliveryCase();
+    expect(delivery.error).toBeNull();
+    expect(delivery.exitCode).toBe(0);
+  });
+
+  test("notifier delivery immediately followed by healthy live serve stays durable across repeated runs (WM-402)", async () => {
+    for (let iteration = 0; iteration < 3; iteration += 1) {
+      const { delivery } = await runNotifierDeliveryCase();
+      expect(delivery.error).toBeNull();
+      await assertHealthyLiveServe();
+    }
+  });
+
+  test("notifier cleanup quiesces a failed pending delivery before closing SQLite (WM-402)", async () => {
+    await expect(
+      runNotifierDeliveryCase({ failWhilePending: true }),
+    ).rejects.toThrow(
+      "intentional assertion failure while notifier delivery is pending",
+    );
+    await assertHealthyLiveServe();
+  });
+});
